Split main module
