+++ whitelist.txt
@@ -9,24 +9,17 @@ fmt
 intersphinx
 intervall
 lk
-lock1
-lock2
 lockfile
-merchantability
 msvcrt
 nblck
 nitpicky
-noninfringement
 param
 pygments
 rdwr
-returnproxy
-seealso
 skipif
 src
 tmp
 trunc
 typehints
 unlck
-unlicense
 wronly
